pyproject.toml
[tool.poetry]
name = "ytkratkobot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
python-dotenv = "^1.0.1"
pydantic = "2.5.3"
requests = "^2.31.0"
httpx = "^0.27.0"
selenium = "^4.19.0"
webdriver-manager = "^4.0.1"
telegraph = {extras = ["aio"], version = "^2.2.0"}
teratts = "^1.0"
yt-dlp = "^2024.4.9"

[tool.poetry.group.dev.dependencies]
psutil = "^5.9.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
